requirements.txt
aiohttp
discord.py
python-dotenv
requests
aiodns
sequence